Merged johnny_work and Joseph-work branches into here.
Corrected where when you put in a letter at the end, it will hide the keyboard.

Added that when you press the backspace button on the keyboard, it will go back to the last text field and remove it on top of the text field you are on (still has some bugs to where when you type in all letters, tapping a text field might now make the backspace work)

Added message to how many points and coins you get when you complete a word.

Change a bit of the UI for the reward message to make it look more appealing.

Added users points to the main screen for easy viewing of their points.